rename param_square_solvers -> implicit_functions
