+++ pyomo/contrib/pynumero/algorithms/solvers/implicit_functions.py
